Change spinlock to limited spin lock
This code sometimes deadlocks; to prevent this spinlock for a while
(10240 loops) but then just assume we can try again the next time
the timer is triggered

This prevents our kernel process from being killed by a watchdog; thus causing
lulzactive to discontinue updating the CPU speed. either causing the system
to be stuck at a low speed (sluggish) or enter a rare SoD state.